fix(vault): util functions
